+++ app.js
@@ -24,7 +24,7 @@ app.use((req, res, next) => {
 
 });
 
-app.use((err, req, rex, next) => {
+app.use((err, req, res, next) => {
     res.status(err.status || 500);
     res.json({
         message: "This Does Not Work",
