+++ routes/index.js
@@ -3,7 +3,7 @@
 const { Router } = require('express');
 const router = Router();
 
-router.use(require('./customers'));
+router.use(require('./users-route'));
 router.use(require('./departments'));
 router.use(require('./computers'));
 router.use(require('./orders'));
@@ -15,7 +15,7 @@ router.use(require('./training-program'));
 
 router.get('/', (req, res) => {
     res.json({
-        "customers": "api/v1/customers",
+        "users-route": "api/v1/users-route",
         "departments": "api/v1/departments",
         "computers": "api/v1/computers",
         "orders": "api/v1/orders",
